Refactor Instrumenters into InstrumentationModules - K, L, M
