+++ instrumentation/khttp-0.1/src/main/java/io/opentelemetry/javaagent/instrumentation/khttp/KHttpInstrumentationModule.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.khttp;
+
+import static io.opentelemetry.javaagent.tooling.ClassLoaderMatcher.hasClassesNamed;
+import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.AgentElementMatchers.safeHasSuperType;
+import static java.util.Collections.singletonList;
+import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
+import static net.bytebuddy.matcher.ElementMatchers.returns;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+
+import com.google.auto.service.AutoService;
+import io.opentelemetry.javaagent.tooling.InstrumentationModule;
+import io.opentelemetry.javaagent.tooling.TypeInstrumentation;
+import java.util.List;
+import java.util.Map;
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+@AutoService(InstrumentationModule.class)
+public class KHttpInstrumentationModule extends InstrumentationModule {
+
+  public KHttpInstrumentationModule() {
+    super("khttp");
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      packageName + ".KHttpHeadersInjectAdapter",
+      packageName + ".KHttpTracer",
+      packageName + ".RequestWrapper",
+    };
+  }
+
+  @Override
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return singletonList(new KHttpInstrumentation());
+  }
+
+  private static final class KHttpInstrumentation implements TypeInstrumentation {
+    @Override
+    public ElementMatcher<ClassLoader> classLoaderMatcher() {
+      return hasClassesNamed("khttp.KHttp");
+    }
+
+    @Override
+    public ElementMatcher<TypeDescription> typeMatcher() {
+      return safeHasSuperType(named("khttp.KHttp"));
+    }
+
+    @Override
+    public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+      return singletonMap(
+          isMethod()
+              .and(not(isAbstract()))
+              .and(named("request"))
+              .and(takesArgument(0, named("java.lang.String")))
+              .and(takesArgument(1, named("java.lang.String")))
+              .and(takesArgument(2, named("java.util.Map")))
+              .and(returns(named("khttp.responses.Response"))),
+          KHttpAdvice.class.getName());
+    }
+  }
+}
