+++ instrumentation/lettuce/lettuce-5.0/src/main/java/io/opentelemetry/javaagent/instrumentation/lettuce/v5_0/LettuceInstrumentationModule.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.lettuce.v5_0;
+
+import static io.opentelemetry.javaagent.tooling.ClassLoaderMatcher.hasClassesNamed;
+import static java.util.Arrays.asList;
+import static net.bytebuddy.matcher.ElementMatchers.not;
+
+import com.google.auto.service.AutoService;
+import io.opentelemetry.javaagent.tooling.InstrumentationModule;
+import io.opentelemetry.javaagent.tooling.TypeInstrumentation;
+import java.util.List;
+import net.bytebuddy.matcher.ElementMatcher;
+
+@AutoService(InstrumentationModule.class)
+public class LettuceInstrumentationModule extends InstrumentationModule {
+  public LettuceInstrumentationModule() {
+    super("lettuce", "lettuce-5");
+  }
+
+  @Override
+  public ElementMatcher.Junction<ClassLoader> classLoaderMatcher() {
+    return not(hasClassesNamed("io.lettuce.core.tracing.Tracing"));
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      "io.opentelemetry.javaagent.instrumentation.lettuce.LettuceArgSplitter",
+      packageName + ".LettuceAbstractDatabaseClientTracer",
+      packageName + ".LettuceConnectionDatabaseClientTracer",
+      packageName + ".LettuceDatabaseClientTracer",
+      packageName + ".LettuceInstrumentationUtil",
+      packageName + ".LettuceAsyncBiFunction",
+      packageName + ".rx.LettuceMonoCreationAdvice",
+      packageName + ".rx.LettuceMonoDualConsumer",
+      packageName + ".rx.LettuceFluxCreationAdvice",
+      packageName + ".rx.LettuceFluxTerminationRunnable",
+      packageName + ".rx.LettuceFluxTerminationRunnable$FluxOnSubscribeConsumer"
+    };
+  }
+
+  @Override
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return asList(
+        new LettuceAsyncCommandsInstrumentation(),
+        new LettuceClientInstrumentation(),
+        new LettuceReactiveCommandsInstrumentation());
+  }
+}
