+++ instrumentation/mongo/mongo-3.1/src/main/java/io/opentelemetry/javaagent/instrumentation/mongo/v3_1/MongoClientInstrumentationModule.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.mongo.v3_1;
 
+import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.declaresMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
@@ -16,7 +17,8 @@ import com.google.auto.service.AutoService;
 import com.mongodb.MongoClientOptions;
 import com.mongodb.event.CommandListener;
 import io.opentelemetry.javaagent.instrumentation.mongo.TracingCommandListener;
-import io.opentelemetry.javaagent.tooling.Instrumenter;
+import io.opentelemetry.javaagent.tooling.InstrumentationModule;
+import io.opentelemetry.javaagent.tooling.TypeInstrumentation;
 import java.lang.reflect.Modifier;
 import java.util.Collections;
 import java.util.List;
@@ -26,27 +28,11 @@ import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-@AutoService(Instrumenter.class)
-public final class MongoClientInstrumentation extends Instrumenter.Default {
+@AutoService(InstrumentationModule.class)
+public final class MongoClientInstrumentationModule extends InstrumentationModule {
 
-  public MongoClientInstrumentation() {
+  public MongoClientInstrumentationModule() {
     super("mongo");
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("com.mongodb.MongoClientOptions$Builder")
-        .and(
-            declaresMethod(
-                named("addCommandListener")
-                    .and(
-                        takesArguments(
-                            new TypeDescription.Latent(
-                                "com.mongodb.event.CommandListener",
-                                Modifier.PUBLIC,
-                                null,
-                                Collections.<TypeDescription.Generic>emptyList())))
-                    .and(isPublic())));
   }
 
   @Override
@@ -58,10 +44,34 @@ public final class MongoClientInstrumentation extends Instrumenter.Default {
   }
 
   @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        isMethod().and(isPublic()).and(named("build")).and(takesArguments(0)),
-        MongoClientInstrumentation.class.getName() + "$MongoClientAdvice");
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return singletonList(new MongoClientOptionsBuilderInstrumentation());
+  }
+
+  private static final class MongoClientOptionsBuilderInstrumentation
+      implements TypeInstrumentation {
+    @Override
+    public ElementMatcher<TypeDescription> typeMatcher() {
+      return named("com.mongodb.MongoClientOptions$Builder")
+          .and(
+              declaresMethod(
+                  named("addCommandListener")
+                      .and(
+                          takesArguments(
+                              new TypeDescription.Latent(
+                                  "com.mongodb.event.CommandListener",
+                                  Modifier.PUBLIC,
+                                  null,
+                                  Collections.<TypeDescription.Generic>emptyList())))
+                      .and(isPublic())));
+    }
+
+    @Override
+    public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+      return singletonMap(
+          isMethod().and(isPublic()).and(named("build")).and(takesArguments(0)),
+          MongoClientInstrumentationModule.class.getName() + "$MongoClientAdvice");
+    }
   }
 
   public static class MongoClientAdvice {
