+++ instrumentation/mongo/mongo-async-3.3/src/main/java/io/opentelemetry/javaagent/instrumentation/mongoasync/v3_3/MongoAsyncClientInstrumentationModule.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.mongoasync.v3_3;
 
+import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.declaresField;
 import static net.bytebuddy.matcher.ElementMatchers.declaresMethod;
@@ -17,7 +18,8 @@ import com.google.auto.service.AutoService;
 import com.mongodb.async.client.MongoClientSettings;
 import com.mongodb.event.CommandListener;
 import io.opentelemetry.javaagent.instrumentation.mongo.TracingCommandListener;
-import io.opentelemetry.javaagent.tooling.Instrumenter;
+import io.opentelemetry.javaagent.tooling.InstrumentationModule;
+import io.opentelemetry.javaagent.tooling.TypeInstrumentation;
 import java.lang.reflect.Modifier;
 import java.util.Collections;
 import java.util.List;
@@ -27,28 +29,11 @@ import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-@AutoService(Instrumenter.class)
-public final class MongoAsyncClientInstrumentation extends Instrumenter.Default {
+@AutoService(InstrumentationModule.class)
+public final class MongoAsyncClientInstrumentationModule extends InstrumentationModule {
 
-  public MongoAsyncClientInstrumentation() {
+  public MongoAsyncClientInstrumentationModule() {
     super("mongo");
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("com.mongodb.async.client.MongoClientSettings$Builder")
-        .and(
-            declaresMethod(
-                named("addCommandListener")
-                    .and(
-                        takesArguments(
-                            new TypeDescription.Latent(
-                                "com.mongodb.event.CommandListener",
-                                Modifier.PUBLIC,
-                                null,
-                                Collections.<TypeDescription.Generic>emptyList())))
-                    .and(isPublic())))
-        .and(declaresField(named("commandListeners")));
   }
 
   @Override
@@ -60,10 +45,35 @@ public final class MongoAsyncClientInstrumentation extends Instrumenter.Default 
   }
 
   @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        isMethod().and(isPublic()).and(named("build")).and(takesArguments(0)),
-        MongoAsyncClientInstrumentation.class.getName() + "$MongoAsyncClientAdvice");
+  public List<TypeInstrumentation> typeInstrumentations() {
+    return singletonList(new MongoClientSettingsBuildersInstrumentation());
+  }
+
+  private static final class MongoClientSettingsBuildersInstrumentation
+      implements TypeInstrumentation {
+    @Override
+    public ElementMatcher<TypeDescription> typeMatcher() {
+      return named("com.mongodb.async.client.MongoClientSettings$Builder")
+          .and(
+              declaresMethod(
+                  named("addCommandListener")
+                      .and(
+                          takesArguments(
+                              new TypeDescription.Latent(
+                                  "com.mongodb.event.CommandListener",
+                                  Modifier.PUBLIC,
+                                  null,
+                                  Collections.<TypeDescription.Generic>emptyList())))
+                      .and(isPublic())))
+          .and(declaresField(named("commandListeners")));
+    }
+
+    @Override
+    public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+      return singletonMap(
+          isMethod().and(isPublic()).and(named("build")).and(takesArguments(0)),
+          MongoAsyncClientInstrumentationModule.class.getName() + "$MongoAsyncClientAdvice");
+    }
   }
 
   public static class MongoAsyncClientAdvice {
